Filter Scala-related noise from JaCoCo results
Adds a new "format", `ScalaHTMLReport` that hides instruction and
branch coverage from the tabular views. These are too noisy primarily because
of synthetic null checks that scalac adds around calls to `==` and `##`
and in the constructors of classes with `$outer` parameters.

Pattern matching is also likely to generate more branches than are tested
in many cases, so its not clear if this will be an easy feature to reenable.

Install a custom `ClassAnalyzer` that omits some methods from
   the reports, namely:
     - simple accessors
     - trait forwarders
     - static forwarders
     - extension method forwarders
     - implicit class factories
     - generated case class and AnyVal methods

This is all something of a stopgap solution until JaCoCo offers a richer
API for filtering, which is being discussed for Q1 2014.